moving testUtils to test resources
