+++ src/test/TestUtils/TestUtilsUnitTests.java
@@ -6,6 +6,7 @@ import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoDatabase;
 import com.mongodb.client.model.Filters;
 import org.junit.Test;
+import resources.TestUtils;
 
 import java.util.Objects;
