+++ src/test/UserTest/ChangeAccountSettingsIntegrationTests.java
@@ -3,7 +3,7 @@ package UserTest;
 import Config.MongoConfig;
 import Security.AccountSecurityController;
 import Security.SecurityUtils;
-import TestUtils.TestUtils;
+import resources.TestUtils;
 import User.User;
 import com.mongodb.client.MongoClient;
 import com.mongodb.client.MongoCollection;
