+++ src/test/UserTest/ChangeTwoFactorSettingIntegrationTests.java
@@ -2,9 +2,8 @@ package UserTest;
 
 import Config.MongoConfig;
 import Security.AccountSecurityController;
-import TestUtils.TestUtils;
+import resources.TestUtils;
 import User.User;
-import User.UserMessage;
 import com.mongodb.client.MongoClient;
 import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoDatabase;
